# worklink-api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "APIs For Thrift Store",
        "version": "1.0.0"
    },
    "paths": {
        "/api/usuarios": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get all users",
                "operationId": "getAllUsers",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": []
            }
        },
        "/api/trabalhos": {
            "get": {
                "tags": [
                    "Works"
                ],
                "summary": "Get all works",
                "operationId": "getAllWorks",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/api/trabalhos/{type}": {
            "get": {
                "tags": [
                    "Works"
                ],
                "summary": "Get works by type",
                "operationId": "getWorksByType",
                "parameters": [
                    {
                        "name": "type",
                        "in": "query",
                        "description": "Work type",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/api/trabalho": {
            "post": {
                "tags": [
                    "Trabalho"
                ],
                "summary": "Cria um novo trabalho",
                "description": "Cria um novo trabalho com base nos parâmetros fornecidos",
                "operationId": "postNewWork",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id_usuario": {
                                        "description": "ID do usuário",
                                        "type": "integer"
                                    },
                                    "id_tipo_trabalho": {
                                        "description": "ID do tipo de trabalho",
                                        "type": "integer"
                                    },
                                    "valor": {
                                        "description": "Valor do trabalho",
                                        "type": "number"
                                    },
                                    "data_inicio": {
                                        "description": "Data de início do trabalho",
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "data_fim": {
                                        "description": "Data de término do trabalho",
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "pagamento": {
                                        "description": "Método de pagamento",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "description": "Dados do novo trabalho",
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "errors": {
                                            "description": "Erros de validação",
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "error": {
                                            "description": "Mensagem de erro",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}